mandelbrot-set-python/requirements.txt
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
